third_party/blink/web_tests/platform/mac-mac11-arm64/external/wpt/html/cross-origin-embedder-policy/credentialless/service-worker-coep-none-proxy.https.window-expected.txt

This is a testharness.js-based test.
FAIL COEP:unsafe-none ServiceWorker assert_equals: expected "ServiceWorker: Proxying" but got "ServiceWorker: Fetch success"
Harness: the test ran to completion.
